#255 Made hotkey work again
The problem was that isPressed can be only called once in a single onKeyInput. If it's called, then the isPressed will return false, even though there was no new cycle. This can be fixed with isKeyDown, but it results in fast executions. Mixing the two only slows down for some reason, but it works now